springboot-backend/src/main/java/tech/trestlework/backend/service/EmployeeServiceImpl.java
package tech.trestlework.backend.service;


import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import tech.trestlework.backend.dto.EmployeeDto;
import tech.trestlework.backend.entity.Employee;
import tech.trestlework.backend.exception.ResourceNotFoundException;
import tech.trestlework.backend.mapper.EmployeeMapper;
import tech.trestlework.backend.repository.EmployeeRepository;
import lombok.AllArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.ArrayList;
import java.util.List;
import java.util.stream.Collectors;


@Service
@AllArgsConstructor
public class EmployeeServiceImpl implements EmployeeService {

    private EmployeeRepository employeeRepository;

    @Override
    public EmployeeDto createEmployee(EmployeeDto employeeDto) {

        Employee employee = EmployeeMapper.mapToEmployee(employeeDto);
        Employee savedEmployee = employeeRepository.save(employee);
        return EmployeeMapper.mapToEmployeeDto(savedEmployee);
    }

    @Override
    public EmployeeDto getEmployeeById(Long employeeId) {
        Employee employee = employeeRepository.findById(employeeId)
                .orElseThrow(()->
                        new ResourceNotFoundException("Employee not exists with the given id : "+employeeId));
        return EmployeeMapper.mapToEmployeeDto(employee);

    }

    @Override
    public List<EmployeeDto> getAllEmployees() {
        List<Employee> employees = employeeRepository.findAll();
        return employees.stream().map((employee)->EmployeeMapper.mapToEmployeeDto(employee))
                .collect(Collectors.toList());
    }

    @Override
    public EmployeeDto updateEmployee(Long employeeId, EmployeeDto updatedEmployee) {
        Employee employee = employeeRepository.findById(employeeId).orElseThrow(
                ()->new ResourceNotFoundException("Employee doesn't exist with the given id :" +employeeId)
        );

        employee.setFirstName(updatedEmployee.getFirstName());
        employee.setLastName(updatedEmployee.getLastName());
        employee.setEmailId(updatedEmployee.getEmailId());

        Employee updatedEmployeeObj = employeeRepository.save(employee);
        return EmployeeMapper.mapToEmployeeDto(updatedEmployeeObj);
    }

    @Override
    public void deleteEmployeeById(Long employeeId) {

        employeeRepository.findById(employeeId).orElseThrow(
            () -> new ResourceNotFoundException("Employee doesn't exists with the given id: "+employeeId)
        );
        employeeRepository.deleteById(employeeId);
    }

//    @Override
//    public Page<EmployeeDto> getAllEmployees(Pageable pageable) {
//        Page<Employee> employeesPage = employeeRepository.findAll(pageable);
//        return employeesPage.map(EmployeeMapper::mapToEmployeeDto);
//    }

//    @Override
//    public List<EmployeeDto> findEmployeeWithSorting(String field) {
//        return List.of();
//    }
//
//    @Override
//    public List<EmployeeDto> findEmployeeWithPagination(int offset, int pageSize) {
//        return List.of();
//    }


}